server-requirements.txt
# Server dependencies for Parallel Agents
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0

# Client dependencies  
requests==2.31.0
websocket-client==1.6.4

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1 